Day14/requirements.txt
pandas
requests
python-dotenv
matplotlib
supabase
